// src/libs/api/GetWeather.ts
import { WEATHER_API_KEY, WEATHER_API_URL } from ".";

export default async function GetWeather(
  city: string
): Promise<GetWeatherProps> {
  const request = await fetch(
    `${WEATHER_API_URL}/weather?q=${city}&units=Metric&appid=${WEATHER_API_KEY}`
  );
  if (!request.ok) {
    throw Error("Failed fetch forecest");
  }
  const respond = (await request.json()) as unknown;
  return respond as GetWeatherProps;
}

// Generated by https://quicktype.io
// ctrl shift p
// paste JSON as type
export interface GetWeatherProps {
  coord: Coord;
  weather: Weather[];
  base: string;
  main: Main;
  visibility: number;
  wind: Wind;
  rain: Rain;
  clouds: Clouds;
  dt: number;
  sys: Sys;
  timezone: number;
  id: number;
  name: string;
  cod: number;
}

export interface Clouds {
  all: number;
}

export interface Coord {
  lon: number;
  lat: number;
}

export interface Main {
  temp: number;
  feels_like: number;
  temp_min: number;
  temp_max: number;
  pressure: number;
  humidity: number;
}

export interface Rain {
  "1h": number;
}

export interface Sys {
  type: number;
  id: number;
  country: string;
  sunrise: number;
  sunset: number;
}

export interface Weather {
  id: number;
  main: string;
  description: string;
  icon: string;
}

export interface Wind {
  speed: number;
  deg: number;
}